Cook the dishes in the kitchen.
Overview: 

You are given a skeleton project called Homework3-template, whose file structure is shown
in Figure 1. The project is incomplete in its current form, and would run correctly only when
Kitchen.java and Order.java are properly implemented. The goal of this homework is thus to
appropriately design the classes to fix the problem.